requirements.txt
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
streamlit>=1.33.0,<2.0.0
plotly>=5.20.0,<6.0.0
ccxt>=4.3.0,<5.0.0
requests>=2.31.0
python-dotenv>=1.0.0,<2.0.0
yfinance>=0.2.0,<1.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# NEW FEATURES - Added Oct 2025
flask>=2.3.0          # TradingView webhook server
openpyxl>=3.1.0       # Excel export for trading journal
pybit>=5.6.0          # Bybit v5 API (if not using ccxt)
flask-socketio>=5.3.6
eventlet>=0.33.0
# pandas-ta removed - incompatible with Python 3.14 (using manual implementations instead)